more sonarqube issue fixes

+++ apps/bfd-server-ng/src/main/java/gov/cms/bfd/server/ng/claim/model/ClaimLineRxDaysSupplyQuantity.java
@@ -16,6 +16,6 @@ class ClaimLineRxDaysSupplyQuantity {
     return supportingInfoFactory
         .createSupportingInfo()
         .setCategory(CarinSupportingInfoCategory.DAYS_SUPPLY.toFhir())
-        .setValue(new Quantity().setValue((long) daysSupply).setUnit("days"));
+        .setValue(new Quantity().setValue(daysSupply).setUnit("days"));
   }
 }
